refactor: enum 클래스 분리 (#190)
