--- src/main/java/com/tebutebu/apiserver/domain/CommentType.java
@@ -1,7 +1,0 @@
-package com.tebutebu.apiserver.domain;
-
-public enum CommentType {
-
-    USER, AI
-
-}
