+++ src/main/java/com/tebutebu/apiserver/domain/converter/MemberRoleConverter.java
@@ -1,6 +1,6 @@
 package com.tebutebu.apiserver.domain.converter;
 
-import com.tebutebu.apiserver.domain.MemberRole;
+import com.tebutebu.apiserver.domain.enums.MemberRole;
 import jakarta.persistence.Converter;
 
 @Converter(autoApply = true)
