+++ src/main/java/com/tebutebu/apiserver/domain/enums/CommentType.java
@@ -1,0 +1,7 @@
+package com.tebutebu.apiserver.domain.enums;
+
+public enum CommentType {
+
+    USER, AI
+
+}
