+++ src/main/java/com/tebutebu/apiserver/domain/enums/Course.java
@@ -1,4 +1,4 @@
-package com.tebutebu.apiserver.domain;
+package com.tebutebu.apiserver.domain.enums;
 
 public enum Course {
