+++ src/main/java/com/tebutebu/apiserver/domain/enums/MemberRole.java
@@ -1,4 +1,4 @@
-package com.tebutebu.apiserver.domain;
+package com.tebutebu.apiserver.domain.enums;
 
 public enum MemberRole {
